feat: Add prettier-plugin-astro dependency
This commit adds the prettier-plugin-astro dependency to the project, with a version of 0.13.0. This plugin is used for formatting Astro code. It is added as a dev dependency and is optional. The prettier-plugin-astro package has dependencies on @astrojs/compiler and sass-formatter. The @astrojs/compiler version 1.8.2 is also added as a devOptional dependency. The prettier-plugin-astro package requires Node.js version 14.15.0 or higher.